+++ Magento/Mtf/Client/Driver/Facebook/Driver.php
@@ -577,13 +577,18 @@ final class Driver implements DriverInterface
     /**
      * Change the focus to a frame in the page by locator
      *
-     * @param Locator $locator
+     * @param Locator|null $locator
      * @return void
      */
-    public function switchToFrame(Locator $locator)
+    public function switchToFrame(Locator $locator = null)
     {
-        $this->eventManager->dispatchEvent(['switch_to_frame'], [(string) $locator]);
-        $this->driver->switchTo()->frame($this->getElement($locator));
+        if (null === $locator) {
+            $this->eventManager->dispatchEvent(['switch_to_frame'], [(string) $locator]);
+            $this->driver->switchTo()->frame($this->getElement($locator));
+        } else {
+            $this->eventManager->dispatchEvent(['switch_to_frame'], ['Switch to main window']);
+            $this->driver->switchTo()->frame();
+        }
     }
 
     /**